Implement a better `force_dpoint` algorithm
We now produce 20x the number of participants and calculate the DP of
the population. Then move the trials around so that the DP of each
condition is exactly the specific value. Finally, we sample the
participants from this population.

Problem: this occupies A TON of memory! REALLY A TON! This *may* cause
some computers to crash =/