Start the code for a GUI.
I think I progressed a lot! I believe I have everything I need to make
the GUI now. Once the GUI is working, I can just "extract" the values
from all the fields of the GUI and populate the variable "params" with
them (from the `run_generator.py` file).

Hopefully it will be straighforward to do all of this now.