Now the code doesn't output those additional files anymore
We decided to remove them because they always referred to information
that needed context to be interpreted, and we were afraid that anyone
(or even us in the future) would end up getting confused with what they
meant. For example, the file `actual_divergence_points.txt` actually
showed the divergence points *before* the shift (when the option to
force the divergence point is set), and we were afraid that people would
see the files and think that they indicated the result of the DPA (i.e.,
that they'd run their own DPA expecting to find those values).